LayoutTests/platform/blackberry/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


input patternMismatch: Pattern mismatch
input valueMissing: Input value missing
textarea valueMissing: Input value missing
select valueMissing: Input value missing
input typeMismatch: Valid email address required
input badInput: Input type mismatch
badInput and valueMissing:
PASS numberInput.validationMessage is nonRequiredBadInputMessage
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS happyOutput.validationMessage is ''
PASS happyObject.validationMessage is ''
PASS happyKeygen.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
